EES-4237 Add option to save the permalink table content
